Rocket landing

+++ MapTools/Passable.java
@@ -4,21 +4,21 @@ import bc.*;
 
 public class Passable{
     private static GameController gc;
-    private static short[][] passabilityMatEarth, passabilityMatMars;
-    private static short passableTotalEarth, passableTotalMars;
+    private static int[][] passabilityMatEarth, passabilityMatMars;
+    private static int passableTotalEarth, passableTotalMars;
 
     public static void setup(GameController gameC){
         gc = gameC;
         passabilityMat(gc.startingMap(Planet.Earth));
         passabilityMat(gc.startingMap(Planet.Mars));
     }
-    public static short[][] matrix(Planet p){
+    public static int[][] matrix(Planet p){
         if(p == Planet.Earth)
             return passabilityMatEarth;
         else
             return passabilityMatMars;
     }
-    public static short total(Planet p){
+    public static int total(Planet p){
         if(p == Planet.Earth)
             return passableTotalEarth;
         else
@@ -26,8 +26,8 @@ public class Passable{
     }
     private static void passabilityMat(PlanetMap pm){
         //same outputs as PlanetMap.isPassableTerrainAt
-        short[][] out = new short[(int)pm.getHeight()][(int)pm.getWidth()];
-        short total = 0;
+        int[][] out = new int[(int)pm.getHeight()][(int)pm.getWidth()];
+        int total = 0;
         for(int y = 0; y < pm.getHeight(); y++){
             for(int x = 0; x < pm.getWidth(); x++){
                 out[y][x] = pm.isPassableTerrainAt(new MapLocation(pm.getPlanet(), x, y));
